increase container lookup performance
Replace _find_container_by_name method with _find_container which
reduces the overhead of calling inspect_container for each running
container. Now in order to find a container we are looking up by the
container name instead of instance name, which helps to reduce the
number of API calls to only one.

Change-Id: I6c7e5573a56231a9da54ac985fb20be3d4ccc849